+++ app/src/main/java/com/pz/activities/ShootingRangeViewModel.java
@@ -39,7 +39,7 @@ public class ShootingRangeViewModel extends AndroidViewModel {
         return mAllCalibers;
     }
 
-    public void updateWeapon(int weapon_id,byte[] weapon_image,String weaponModel,int fk_caliber_id,int priceForShoot){
+    public void updateWeapon(int weapon_id,byte[] weapon_image,String weaponModel,int fk_caliber_id,double priceForShoot){
         mRepository.updateWeapon(weapon_id,weapon_image,weaponModel,fk_caliber_id,priceForShoot);
     }
